requirements.txt
pyqt5
pycryptodome
pyperclip